requirements.txt
pytest-playwright
librosa
pyaudio
numpy
openai
python-dotenv
opencv-python
